requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.4
pandas==1.0.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
requests==2.23.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
SQLAlchemy==1.3.17
threadpoolctl==2.0.0
urllib3==1.25.9
Werkzeug==1.0.1
